change background for tuto in header
also move preloading from App to MainContainer
